Align resource classes with upstream
We are currently seeing an error where there is an 'illegal instruction'
when building some of the go tooling.  It's possible that because we are
building the ci-image on an 'extra large' VM rather than the 'medium'
one of upstream that some of the libs/binaries are using CPU
optimizations that are not consistently available.